[#noissue] Add shared ECharts chart helpers
Extract the pieces shared by the ECharts-based charts into lib/charts:
- useEChartsInstance: init/resize/dispose lifecycle, plus an optional
  connect group for synchronized tooltips.
- echartsLegendLayout: bottom-legend row/grid-bottom calculation (skips the
  legend space when there are no series).
- echartsTimeSeriesFormat: category date-axis label and axis tooltip
  builders, with HTML escaping to prevent injection from series names.
- escapeHTMLEntities string util (+ tests).
Refactor the URL Statistic chart to consume these helpers.
